SAA-62 introduction of activity service for retrieval of a single activity by ID.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/repository/ActivityRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository
 import uk.gov.justice.digital.hmpps.hmppsactivitiesmanagementapi.entity.Activity
 
 @Repository
-interface ActivityRepository : JpaRepository<Activity, Int>
+interface ActivityRepository : JpaRepository<Activity, Long>
